fix: Movement State Machine for Beam
Created the Movement State Machine for the beam. It controls what the linear and angular velocity should be in each state e.g. stationary or rotating.
The _integrate_forces method then gets these values directly from the state machine.
The Basic State Machine now keeps track of the right and left directions which is then fed to the Movement SM. These directions are set differently depending on whether the beam is in reset mode or user input
Minor fix to not trigger the round animation when round is set to 1